+++ nv_flutter_sdk_app/android/app/src/main/kotlin/com/notifyvisitors/flutter/sdkapp/MainApplication.kt
@@ -14,8 +14,8 @@ class MainApplication : Application() {
     override fun onCreate() {
         super.onCreate()
         
-        Log.d("rn-nv-android", "nvBrandID = " + nvBrandID + " nvSecreKey = " + nvSecreKey)
-        NotifyvisitorsPlugin.register(this, nvBrandID, nvSecreKey);
+        Log.i("rn-nv-android", "nvBrandID = " + nvBrandID + " nvSecreKey = " + nvSecreKey)
+        NotifyvisitorsPlugin.register(this, 7577, "DB52A5B00BB0D3BF426639A1B9FCF2F7");
     
     }
 }